Moved tests to properly directory
