+++ src/Component/tests/Symfony/DependencyInjection/Compiler/MetadataMutatorPassTest.php
